Added SC_Chart and SC_timeSeriesResultChart classes to the Common.
Added Two clases to the Simcenter Common. Obe is SC_Chart amnd the Other is SC_TimeSeriesresultChart. These two classes are desigend to facilitate the use of charts and timeseries type result in the Frontend application. With SC_TimeSeriesresultChart class, one can show a chart with a cmbobox for different metrics from a json file. If One sticks to the conventions of timeseries used in REWET Result in producing their JSON file, then they can use this class without a need to program everything all again.